requirements.txt
fastapi>=0.100.0,<1.0
uvicorn>=0.35.0,<1.0
SQLAlchemy>=1.4.23,<2.1.0
psycopg2-binary>=2.9.9,<3
alembic>=1.17.2,<2
python-dotenv>=1.2.1,<2
janus>=2.0.0,<3
opentelemetry-api>=1.39,<2
opentelemetry-sdk>=1.39,<2
opentelemetry-exporter-otlp>=1.39,<2
opentelemetry-instrumentation-fastapi>=0.60b0,<1.0
redis[hiredis]>=5.0,<8
